Super hacky quick fix for code mirror induced copy/paste issues
To get the appropriate context menu for a text area, there is an on contextmenu click event handler. Codemirror blocks the default action, captures the mouse position, temporarily unhides the psuedo-textarea (transparently and not self-evidently to the end user) and attempts to position the textarea element on screen where the click originated. By than continuing the default action, the browser believes the click has been made in a true text area and shows the appropriate context menu. 

However, if you adjust the opacity on line 9615, it becomes apparent that the positioning of that element is not in line with the mouse click and thus the appropriate menu never shows. By making slight adjustments to the top and left style attributes given to the psuedo textarea element when it's shown, we can place it in the appropriate position and thus end up with the intended menu.